TIDIJ-191 Moved files that will be used in all modules to core.
