+++ core/lib/state/vehicle_state.dart
