+++ core/lib/vehicle.dart
