+++ core/pubspec.yaml
@@ -27,7 +27,9 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-
+  fleetcarpooling:
+    path: ../fleetcarpooling
+   
 dev_dependencies:
   flutter_test:
     sdk: flutter
