+++ fleetcarpooling/lib/Modularity/Pages/add_vehicle_manually_form.dart
@@ -1,9 +1,9 @@
+import 'package:core/event/vehicle_event.dart';
+import 'package:core/state/vehicle_state.dart';
+import 'package:core/vehicle.dart';
 import 'package:fleetcarpooling/Modularity/Pages/add_vehicle_interface.dart';
 import 'package:fleetcarpooling/Modularity/bloc/vehicle_bloc.dart';
-import 'package:fleetcarpooling/Modularity/event/vehicle_event.dart';
-import 'package:fleetcarpooling/Modularity/models/vehicle.dart';
 import 'package:fleetcarpooling/Modularity/service/image_service.dart';
-import 'package:fleetcarpooling/Modularity/state/vehicle_state.dart';
 import 'package:fleetcarpooling/ui_elements/buttons.dart';
 import 'package:fleetcarpooling/ui_elements/colors';
 import 'package:flutter/material.dart';
