+++ fleetcarpooling/lib/Modularity/bloc/vehicle_bloc.dart
@@ -1,7 +1,7 @@
 import 'package:bloc/bloc.dart';
-import 'package:fleetcarpooling/Modularity/event/vehicle_event.dart';
+import 'package:core/event/vehicle_event.dart';
+import 'package:core/state/vehicle_state.dart';
 import 'package:fleetcarpooling/Modularity/service/vehicle_service.dart';
-import 'package:fleetcarpooling/Modularity/state/vehicle_state.dart';
 
 class VehicleBloc extends Bloc<VehicleEvent, VehicleState> {
   final VehicleRepository _vehicleRepository;
