+++ fleetcarpooling/lib/pages/admin_home_page.dart
@@ -1,4 +1,4 @@
-import 'package:fleetcarpooling/Modularity/Pages/add_vehicle_selection.dart';
+import 'package:core/add_vehicle_selection.dart';
 import 'package:fleetcarpooling/pages/all_reservations.dart';
 import 'package:fleetcarpooling/pages/all_users_form.dart';
 import 'package:fleetcarpooling/pages/delete_disable_form.dart';
