+++ fleetcarpooling/lib/pages/map.dart
@@ -1,5 +1,5 @@
+import 'package:core/vehicle.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:fleetcarpooling/Modularity/models/vehicle.dart';
 import 'package:fleetcarpooling/ReservationService/reservation_service.dart';
 import 'package:fleetcarpooling/VehicleManagamentService/vehicle_managament_service.dart';
 import 'package:fleetcarpooling/handlers/udp_manager.dart';
