+++ fleetcarpooling/lib/pages/vehicle_management_form.dart
@@ -1,4 +1,4 @@
-import 'package:fleetcarpooling/Modularity/Pages/add_vehicle_selection.dart';
+import 'package:core/add_vehicle_selection.dart';
 import 'package:fleetcarpooling/pages/delete_disable_form.dart';
 import 'package:fleetcarpooling/ui_elements/buttons.dart';
 import 'package:fleetcarpooling/ui_elements/colors';
