+++ fleetcarpooling/pubspec.yaml
@@ -32,7 +32,8 @@ dependencies:
   flutter:
     sdk: flutter
   intl: ^0.18.0
-
+  core:
+    path: ../core
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
